% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{Task}
\alias{Task}
\title{TaskQueue API Objects 
Accesses a Google App Engine Pull Task Queue over REST.}
\usage{
Task(enqueueTimestamp = NULL, id = NULL, leaseTimestamp = NULL,
  payloadBase64 = NULL, queueName = NULL)
}
\arguments{
\item{enqueueTimestamp}{Time (in seconds since the epoch) at which the task was enqueued}

\item{id}{Name of the task}

\item{leaseTimestamp}{Time (in seconds since the epoch) at which the task lease will expire}

\item{payloadBase64}{A bag of bytes which is the task payload}

\item{queueName}{Name of the queue that the task is in}
}
\value{
Task object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:49:12
filename: /Users/mark/dev/R/autoGoogleAPI/googletaskqueuev1beta1.auto/R/taskqueue_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Task Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

