% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_objects.R
\name{TaskQueue.acl}
\alias{TaskQueue.acl}
\title{TaskQueue.acl Object}
\usage{
TaskQueue.acl(adminEmails = NULL, consumerEmails = NULL,
  producerEmails = NULL)
}
\arguments{
\item{adminEmails}{Email addresses of users who are 'admins' of the TaskQueue}

\item{consumerEmails}{Email addresses of users who can 'consume' tasks from the TaskQueue}

\item{producerEmails}{Email addresses of users who can 'produce' tasks into the TaskQueue}
}
\value{
TaskQueue.acl object
}
\description{
TaskQueue.acl Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
ACLs that are applicable to this TaskQueue object.
}
\seealso{
Other TaskQueue functions: \code{\link{TaskQueue.stats}},
  \code{\link{TaskQueue}}
}

