% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_functions.R
\docType{package}
\name{taskqueue_googleAuthR}
\alias{taskqueue_googleAuthR}
\alias{taskqueue_googleAuthR-package}
\title{TaskQueue API
Accesses a Google App Engine Pull Task Queue over REST.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:18:57
filename: /Users/mark/dev/R/autoGoogleAPI/googletaskqueuev1beta1.auto/R/taskqueue_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/taskqueue
\item https://www.googleapis.com/auth/taskqueue.consumer
}
}
