% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskqueue_functions.R
\name{tasks.delete}
\alias{tasks.delete}
\title{Delete a task from a TaskQueue.}
\usage{
tasks.delete(project, taskqueue, task)
}
\arguments{
\item{project}{The project under which the queue lies}

\item{taskqueue}{The taskqueue to delete a task from}

\item{task}{The id of the task to delete}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/taskqueue
\item https://www.googleapis.com/auth/taskqueue.consumer
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/taskqueue, https://www.googleapis.com/auth/taskqueue.consumer)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/appengine/docs/python/taskqueue/rest}{Google Documentation}
}

